backend/requirements.txt
django==5.1.2
djangorestframework==3.15.2
django-environ==0.11.2